requirements.txt
accelerate==1.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
braceexpand==0.1.7
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
decord==0.6.0
docker-pycreds==0.4.0
einops==0.8.0
eval_type_backport==0.2.2
filelock==3.13.1
fonttools==4.55.8
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
iopath==0.1.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
peft==0.13.0
pillow==11.0.0
platformdirs==4.3.6
portalocker==3.1.1
progressbar2==4.5.0
protobuf==5.29.3
psutil==6.1.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-image==0.24.0
scipy==1.13.1
sentencepiece==0.2.0
sentry-sdk==2.20.0
setproctitle==1.3.4
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tifffile==2024.8.30
timm==1.0.14
tokenizers==0.15.2
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
transformers==4.37.0
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
visual-genome==1.1.1
wandb==0.19.6
webdataset==0.2.100
zipp==3.21.0
